Move startpoint test from TestZkLocalApplicationRunner into its own test harness class.
