+++ samza-test/src/test/java/org/apache/samza/test/processor/TestZkLocalApplicationRunner.java
@@ -24,7 +24,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -33,18 +32,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
-import java.util.TreeMap;
 import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 import org.I0Itec.zkclient.ZkClient;
 import org.apache.kafka.clients.admin.NewTopic;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.samza.Partition;
 import org.apache.samza.SamzaException;
 import org.apache.samza.application.TaskApplication;
@@ -71,21 +65,10 @@ import org.apache.samza.metrics.MetricsRegistryMap;
 import org.apache.samza.runtime.ApplicationRunner;
 import org.apache.samza.runtime.ApplicationRunners;
 import org.apache.samza.runtime.LocalApplicationRunner;
-import org.apache.samza.startpoint.Startpoint;
-import org.apache.samza.startpoint.StartpointManager;
-import org.apache.samza.startpoint.StartpointOldest;
-import org.apache.samza.startpoint.StartpointSpecific;
-import org.apache.samza.startpoint.StartpointTimestamp;
-import org.apache.samza.startpoint.StartpointUpcoming;
-import org.apache.samza.system.IncomingMessageEnvelope;
-import org.apache.samza.system.SystemAdmin;
-import org.apache.samza.system.SystemAdmins;
-import org.apache.samza.system.SystemStream;
 import org.apache.samza.system.SystemStreamPartition;
-import org.apache.samza.task.TaskCallback;
 import org.apache.samza.test.StandaloneTestUtils;
 import org.apache.samza.test.harness.IntegrationTestHarness;
-import org.apache.samza.util.CoordinatorStreamUtil;
+import org.apache.samza.test.util.TestKafkaEvent;
 import org.apache.samza.util.NoOpMetricsRegistry;
 import org.apache.samza.util.ReflectionUtil;
 import org.apache.samza.zk.ZkJobCoordinatorFactory;
@@ -121,7 +104,6 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   private static final int NUM_KAFKA_EVENTS = 300;
   private static final int ZK_CONNECTION_TIMEOUT_MS = 5000;
   private static final int ZK_SESSION_TIMEOUT_MS = 10000;
-  private static final int ZK_TEST_PARTITION_COUNT = 5;
   private static final String TEST_SYSTEM = "TestSystemName";
   private static final String TEST_SSP_GROUPER_FACTORY = "org.apache.samza.container.grouper.stream.GroupByPartitionFactory";
   private static final String TEST_TASK_GROUPER_FACTORY = "org.apache.samza.container.grouper.task.GroupByContainerIdsFactory";
@@ -130,12 +112,9 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   private static final String TASK_SHUTDOWN_MS = "10000";
   private static final String JOB_DEBOUNCE_TIME_MS = "10000";
   private static final String BARRIER_TIMEOUT_MS = "10000";
-  private static final String[] PROCESSOR_IDS = new String[] {"0000000000", "0000000001", "0000000002", "0000000003"};
+  private static final String[] PROCESSOR_IDS = new String[] {"0000000000", "0000000001", "0000000002"};
 
-  private String inputKafkaTopic1;
-  private String inputKafkaTopic2;
-  private String inputKafkaTopic3;
-  private String inputKafkaTopic4;
+  private String inputKafkaTopic;
   private String outputKafkaTopic;
   private String inputSinglePartitionKafkaTopic;
   private String outputSinglePartitionKafkaTopic;
@@ -143,7 +122,6 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   private ApplicationConfig applicationConfig1;
   private ApplicationConfig applicationConfig2;
   private ApplicationConfig applicationConfig3;
-  private ApplicationConfig applicationConfig4;
   private String testStreamAppName;
   private String testStreamAppId;
   private MetadataStore zkMetadataStore;
@@ -160,10 +138,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     String uniqueTestId = UUID.randomUUID().toString();
     testStreamAppName = String.format("test-app-name-%s", uniqueTestId);
     testStreamAppId = String.format("test-app-id-%s", uniqueTestId);
-    inputKafkaTopic1 = String.format("test-input-topic1-%s", uniqueTestId);
-    inputKafkaTopic2 = String.format("test-input-topic2-%s", uniqueTestId);
-    inputKafkaTopic3 = String.format("test-input-topic3-%s", uniqueTestId);
-    inputKafkaTopic4 = String.format("test-input-topic4-%s", uniqueTestId);
+    inputKafkaTopic = String.format("test-input-topic-%s", uniqueTestId);
     outputKafkaTopic = String.format("test-output-topic-%s", uniqueTestId);
     inputSinglePartitionKafkaTopic = String.format("test-input-single-partition-topic-%s", uniqueTestId);
     outputSinglePartitionKafkaTopic = String.format("test-output-single-partition-topic-%s", uniqueTestId);
@@ -178,26 +153,20 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     applicationConfig2 = new ApplicationConfig(new MapConfig(configMap));
     configMap.put(JobConfig.PROCESSOR_ID, PROCESSOR_IDS[2]);
     applicationConfig3 = new ApplicationConfig(new MapConfig(configMap));
-    configMap.put(JobConfig.PROCESSOR_ID, PROCESSOR_IDS[3]);
-    applicationConfig4 = new ApplicationConfig(new MapConfig(configMap));
 
     ZkClient zkClient = new ZkClient(zkConnect(), ZK_CONNECTION_TIMEOUT_MS, ZK_CONNECTION_TIMEOUT_MS, new ZkStringSerializer());
     ZkKeyBuilder zkKeyBuilder = new ZkKeyBuilder(ZkJobCoordinatorFactory.getJobCoordinationZkPath(applicationConfig1));
     zkUtils = new ZkUtils(zkKeyBuilder, zkClient, ZK_CONNECTION_TIMEOUT_MS, ZK_SESSION_TIMEOUT_MS, new NoOpMetricsRegistry());
     zkUtils.connect();
 
-    ImmutableMap<String, Integer> topicToPartitionCount = ImmutableMap.<String, Integer>builder()
-        .put(inputSinglePartitionKafkaTopic, 1)
-        .put(outputSinglePartitionKafkaTopic, 1)
-        .put(inputKafkaTopic1, ZK_TEST_PARTITION_COUNT)
-        .put(inputKafkaTopic2, ZK_TEST_PARTITION_COUNT)
-        .put(inputKafkaTopic3, ZK_TEST_PARTITION_COUNT)
-        .put(inputKafkaTopic4, ZK_TEST_PARTITION_COUNT)
-        .put(outputKafkaTopic, ZK_TEST_PARTITION_COUNT)
-        .build();
+    ImmutableMap<String, Integer> topicToPartitionCount = ImmutableMap.of(
+        inputSinglePartitionKafkaTopic, 1,
+        outputSinglePartitionKafkaTopic, 1,
+        inputKafkaTopic, 5,
+        outputKafkaTopic, 5);
 
     List<NewTopic> newTopics =
-        topicToPartitionCount.keySet()
+        ImmutableList.of(inputKafkaTopic, outputKafkaTopic, inputSinglePartitionKafkaTopic, outputSinglePartitionKafkaTopic)
             .stream()
             .map(topic -> new NewTopic(topic, topicToPartitionCount.get(topic), (short) 1))
             .collect(Collectors.toList());
@@ -208,8 +177,8 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
 
   @Override
   public void tearDown() {
-    deleteTopics(ImmutableList.of(inputKafkaTopic1, inputKafkaTopic2, inputKafkaTopic3, inputKafkaTopic4, outputKafkaTopic,
-        inputSinglePartitionKafkaTopic, outputSinglePartitionKafkaTopic));
+    deleteTopics(ImmutableList.of(
+        inputKafkaTopic, outputKafkaTopic, inputSinglePartitionKafkaTopic, outputSinglePartitionKafkaTopic));
     SharedContextFactories.clearAll();
     zkUtils.close();
     super.tearDown();
@@ -219,36 +188,12 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     for (int eventIndex = startIndex; eventIndex < endIndex; eventIndex++) {
       try {
         LOGGER.info("Publish kafka event with index : {} for stream processor: {}.", eventIndex, streamProcessorId);
-        TestStreamApplication.TestKafkaEvent testKafkaEvent =
-            new TestStreamApplication.TestKafkaEvent(streamProcessorId, String.valueOf(eventIndex));
-        producer.send(new ProducerRecord(topic, testKafkaEvent.toString().getBytes()));
+        producer.send(new ProducerRecord(topic, new TestKafkaEvent(streamProcessorId, String.valueOf(eventIndex)).toString().getBytes()));
       } catch (Exception  e) {
         LOGGER.error("Publishing to kafka topic: {} resulted in exception: {}.", new Object[]{topic, e});
         throw new SamzaException(e);
       }
     }
-    producer.flush();
-  }
-
-  // Sends each event synchronously and returns map of event index to event metadata of the sent record/event. Adds the specified delay between sends.
-  private TreeMap<Integer, RecordMetadata> publishKafkaEventsWithDelayPerEvent(String topic, int startIndex, int endIndex, String streamProcessorId, Duration delay) {
-    TreeMap<Integer, RecordMetadata> eventsMetadata = new TreeMap<>();
-    for (int eventIndex = startIndex; eventIndex < endIndex; eventIndex++) {
-      try {
-        LOGGER.info("Publish kafka event with index : {} for stream processor: {}.", eventIndex, streamProcessorId);
-        TestStreamApplication.TestKafkaEvent testKafkaEvent =
-            new TestStreamApplication.TestKafkaEvent(streamProcessorId, String.valueOf(eventIndex));
-        Thread.sleep(delay.toMillis());
-        Future<RecordMetadata> send = producer.send(new ProducerRecord(topic,
-            testKafkaEvent.toString().getBytes()));
-        eventsMetadata.put(eventIndex, send.get(5, TimeUnit.SECONDS));
-      } catch (Exception  e) {
-        LOGGER.error("Publishing to kafka topic: {} resulted in exception: {}.", new Object[]{topic, e});
-        throw new SamzaException(e);
-      }
-    }
-    producer.flush();
-    return eventsMetadata;
   }
 
   private Map<String, String> buildStreamApplicationConfigMap(String appName, String appId, boolean isBatch) {
@@ -386,7 +331,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   @Test
   public void shouldUpdateJobModelWhenNewProcessorJoiningGroupUsingAllSspToSingleTaskGrouperFactory() throws InterruptedException {
     // Set up kafka topics.
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS * 2, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS * 2, PROCESSOR_IDS[0]);
 
     // Configuration, verification variables
     MapConfig testConfig = new MapConfig(ImmutableMap.of(JobConfig.SSP_GROUPER_FACTORY,
@@ -408,7 +353,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch = new CountDownLatch(NUM_KAFKA_EVENTS * 2);
     Config testAppConfig2 = new MapConfig(applicationConfig2, testConfig);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch, null,
         null, testAppConfig2), testAppConfig2);
 
     // Callback handler for appRunner1.
@@ -432,7 +377,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     // Set up stream app appRunner1.
     Config testAppConfig1 = new MapConfig(applicationConfig1, testConfig);
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, null, streamApplicationCallback,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, null, streamApplicationCallback,
         kafkaEventsConsumedLatch, testAppConfig1), testAppConfig1);
     executeRun(appRunner1, testAppConfig1);
 
@@ -477,7 +422,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   @Test
   public void shouldReElectLeaderWhenLeaderDies() throws InterruptedException {
     // Set up kafka topics.
-    publishKafkaEvents(inputKafkaTopic1, 0, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     // Create stream applications.
     CountDownLatch kafkaEventsConsumedLatch = new CountDownLatch(2 * NUM_KAFKA_EVENTS);
@@ -486,13 +431,13 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch3 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
         applicationConfig2), applicationConfig2);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch3, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch3, null, kafkaEventsConsumedLatch,
         applicationConfig3), applicationConfig3);
 
     executeRun(appRunner1, applicationConfig1);
@@ -520,7 +465,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     assertEquals(ApplicationStatus.SuccessfulFinish, appRunner1.status());
 
     kafkaEventsConsumedLatch.await();
-    publishKafkaEvents(inputKafkaTopic1, 0, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     executeRun(appRunner3, applicationConfig3);
     processedMessagesLatch3.await();
@@ -545,7 +490,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
   @Test
   public void shouldFailWhenNewProcessorJoinsWithSameIdAsExistingProcessor() throws InterruptedException {
     // Set up kafka topics.
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     // Create StreamApplications.
     CountDownLatch kafkaEventsConsumedLatch = new CountDownLatch(NUM_KAFKA_EVENTS);
@@ -553,10 +498,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
         applicationConfig2), applicationConfig2);
 
     // Run stream applications.
@@ -568,10 +513,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     processedMessagesLatch2.await();
 
     // Create a stream app with same processor id as SP2 and run it. It should fail.
-    publishKafkaEvents(inputKafkaTopic1, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[2]);
+    publishKafkaEvents(inputKafkaTopic, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[2]);
     kafkaEventsConsumedLatch = new CountDownLatch(NUM_KAFKA_EVENTS);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, null, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, null, null, kafkaEventsConsumedLatch,
         applicationConfig2), applicationConfig2);
     // Fail when the duplicate processor joins.
     expectedException.expect(SamzaException.class);
@@ -586,14 +531,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     }
   }
 
-  public void addMessagesProcessed(List<TestStreamApplication.TestKafkaEvent> messagesProcessed, TestStreamApplication.TestKafkaEvent tke) {
-    messagesProcessed.add(tke);
-  }
-
   @Test
   public void testRollingUpgradeOfStreamApplicationsShouldGenerateSameJobModel() throws Exception {
     // Set up kafka topics.
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap = buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId);
 
@@ -603,9 +544,8 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     configMap.put(JobConfig.PROCESSOR_ID, PROCESSOR_IDS[1]);
     Config applicationConfig2 = new MapConfig(configMap);
 
-    List<TestStreamApplication.TestKafkaEvent> messagesProcessed = new ArrayList<>();
-    TestStreamApplication.StreamApplicationCallback streamApplicationCallback =
-        (TestStreamApplication.TestKafkaEvent tke) -> addMessagesProcessed(messagesProcessed, tke);
+    List<TestKafkaEvent> messagesProcessed = new ArrayList<>();
+    TestStreamApplication.StreamApplicationCallback streamApplicationCallback = messagesProcessed::add;
 
     // Create StreamApplication from configuration.
     CountDownLatch kafkaEventsConsumedLatch = new CountDownLatch(NUM_KAFKA_EVENTS);
@@ -613,10 +553,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, streamApplicationCallback, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
         applicationConfig2), applicationConfig2);
 
     // Run stream application.
@@ -634,7 +574,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     appRunner1.waitForFinish();
 
     int lastProcessedMessageId = -1;
-    for (TestStreamApplication.TestKafkaEvent message : messagesProcessed) {
+    for (TestKafkaEvent message : messagesProcessed) {
       lastProcessedMessageId = Math.max(lastProcessedMessageId, Integer.parseInt(message.getEventData()));
     }
     messagesProcessed.clear();
@@ -642,9 +582,9 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     assertEquals(ApplicationStatus.SuccessfulFinish, appRunner1.status());
 
     processedMessagesLatch1 = new CountDownLatch(1);
-    publishKafkaEvents(inputKafkaTopic1, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
         applicationConfig1), applicationConfig1);
     executeRun(appRunner3, applicationConfig1);
 
@@ -667,7 +607,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
 
   @Test
   public void testShouldStopStreamApplicationWhenShutdownTimeOutIsLessThanContainerShutdownTime() throws Exception {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap = buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId);
     configMap.put(TaskConfig.TASK_SHUTDOWN_MS, "0");
@@ -684,10 +624,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, kafkaEventsConsumedLatch,
         applicationConfig2), applicationConfig2);
 
     executeRun(appRunner1, applicationConfig1);
@@ -709,11 +649,11 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch3 = new CountDownLatch(1);
 
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch3, null, kafkaEventsConsumedLatch,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch3, null, kafkaEventsConsumedLatch,
         applicationConfig3), applicationConfig3);
     executeRun(appRunner3, applicationConfig3);
 
-    publishKafkaEvents(inputKafkaTopic1, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, NUM_KAFKA_EVENTS, 2 * NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     processedMessagesLatch3.await();
     appRunner1.waitForFinish();
@@ -739,14 +679,14 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
    */
   @Test
   public void testShouldGenerateJobModelOnPartitionCountChange() throws Exception {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     // Create StreamApplication from configuration.
     CountDownLatch kafkaEventsConsumedLatch1 = new CountDownLatch(NUM_KAFKA_EVENTS);
     CountDownLatch processedMessagesLatch1 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-            TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch1,
+            TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, kafkaEventsConsumedLatch1,
             applicationConfig1), applicationConfig1);
 
     executeRun(appRunner1, applicationConfig1);
@@ -760,7 +700,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     Assert.assertEquals(5, ssps.size());
 
     // Increase the partition count of input kafka topic to 100.
-    increasePartitionsTo(inputKafkaTopic1, 100);
+    increasePartitionsTo(inputKafkaTopic, 100);
 
     long jobModelWaitTimeInMillis = 10;
     while (Objects.equals(zkUtils.getJobModelVersion(), jobModelVersion)) {
@@ -1002,14 +942,14 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
 
   @Test
   public void testApplicationShutdownShouldBeIndependentOfPerMessageProcessingTime() throws Exception {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     // Create a TaskApplication with only one task per container.
     // The task does not invokes taskCallback.complete for any of the dispatched message.
     CountDownLatch shutdownLatch = new CountDownLatch(1);
     CountDownLatch processedMessagesLatch1 = new CountDownLatch(1);
 
-    TaskApplication taskApplication = new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic1, outputKafkaTopic, processedMessagesLatch1, shutdownLatch, null);
+    TaskApplication taskApplication = new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic, outputKafkaTopic, processedMessagesLatch1, shutdownLatch);
     MapConfig taskApplicationConfig = new MapConfig(ImmutableList.of(applicationConfig1,
         ImmutableMap.of(TaskConfig.MAX_CONCURRENCY, "1", JobConfig.SSP_GROUPER_FACTORY, "org.apache.samza.container.grouper.stream.AllSspToSingleTaskGrouperFactory")));
     ApplicationRunner appRunner = ApplicationRunners.getApplicationRunner(taskApplication, taskApplicationConfig);
@@ -1039,7 +979,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
    */
   @Test
   public void testAgreeingOnSameRunIdForBatch() throws InterruptedException {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap =
         buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId, true);
@@ -1054,10 +994,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, null,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, null,
         applicationConfig2), applicationConfig2);
 
     executeRun(appRunner1, applicationConfig1);
@@ -1091,7 +1031,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
    */
   @Test
   public void testNewProcessorGetsSameRunIdForBatch() throws InterruptedException {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap =
         buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId, true);
@@ -1107,10 +1047,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, null,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, null,
         applicationConfig2), applicationConfig2);
 
     executeRun(appRunner1, applicationConfig1);
@@ -1130,7 +1070,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     //Bring up a new processsor
     CountDownLatch processedMessagesLatch3 = new CountDownLatch(1);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch3, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch3, null, null,
         applicationConfig3), applicationConfig3);
     executeRun(appRunner3, applicationConfig3);
     processedMessagesLatch3.await();
@@ -1161,7 +1101,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
    */
   @Test
   public void testAllProcesssorDieNewProcessorGetsNewRunIdForBatch() throws InterruptedException {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap =
         buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId, true);
@@ -1177,10 +1117,10 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
 
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, null,
         applicationConfig1), applicationConfig1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, null,
         applicationConfig2), applicationConfig2);
 
     executeRun(appRunner1, applicationConfig1);
@@ -1208,7 +1148,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     //Bring up a new processsor
     CountDownLatch processedMessagesLatch3 = new CountDownLatch(1);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch3, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch3, null, null,
         applicationConfig3), applicationConfig3);
     executeRun(appRunner3, applicationConfig3);
     processedMessagesLatch3.await();
@@ -1235,7 +1175,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
    */
   @Test
   public void testFirstProcessorDiesButSameRunIdForBatch() throws InterruptedException {
-    publishKafkaEvents(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
+    publishKafkaEvents(inputKafkaTopic, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0]);
 
     Map<String, String> configMap =
         buildStreamApplicationConfigMap(testStreamAppName, testStreamAppId, true);
@@ -1248,7 +1188,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
 
     CountDownLatch processedMessagesLatch1 = new CountDownLatch(1);
     ApplicationRunner appRunner1 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch1, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch1, null, null,
         applicationConfig1), applicationConfig1);
     executeRun(appRunner1, applicationConfig1);
 
@@ -1260,7 +1200,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     // bring up second processor
     CountDownLatch processedMessagesLatch2 = new CountDownLatch(1);
     ApplicationRunner appRunner2 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch2, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch2, null, null,
         applicationConfig2), applicationConfig2);
     executeRun(appRunner2, applicationConfig2);
 
@@ -1277,7 +1217,7 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     //Bring up a new processsor
     CountDownLatch processedMessagesLatch3 = new CountDownLatch(1);
     ApplicationRunner appRunner3 = ApplicationRunners.getApplicationRunner(TestStreamApplication.getInstance(
-        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic1), outputKafkaTopic, processedMessagesLatch3, null, null,
+        TEST_SYSTEM, ImmutableList.of(inputKafkaTopic), outputKafkaTopic, processedMessagesLatch3, null, null,
         applicationConfig3), applicationConfig3);
     executeRun(appRunner3, applicationConfig3);
     processedMessagesLatch3.await();
@@ -1292,159 +1232,6 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
     appRunner2.waitForFinish();
     appRunner3.kill();
     appRunner3.waitForFinish();
-  }
-
-  private void writeStartpoints(StartpointManager startpointManager, String streamName, Integer partitionCount, Startpoint startpoint) {
-    for (int p = 0; p < partitionCount; p++) {
-      startpointManager.writeStartpoint(new SystemStreamPartition(TEST_SYSTEM, streamName, new Partition(p)), startpoint);
-    }
-  }
-
-  @Test
-  public void testStartpoints() throws InterruptedException {
-    TreeMap<Integer, RecordMetadata> sentEvents1 =
-        publishKafkaEventsWithDelayPerEvent(inputKafkaTopic1, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[0], Duration.ofMillis(2));
-    TreeMap<Integer, RecordMetadata> sentEvents2 =
-        publishKafkaEventsWithDelayPerEvent(inputKafkaTopic2, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[1], Duration.ofMillis(2));
-    TreeMap<Integer, RecordMetadata> sentEvents3 =
-        publishKafkaEventsWithDelayPerEvent(inputKafkaTopic3, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[2], Duration.ofMillis(2));
-    TreeMap<Integer, RecordMetadata> sentEvents4 =
-        publishKafkaEventsWithDelayPerEvent(inputKafkaTopic4, 0, NUM_KAFKA_EVENTS, PROCESSOR_IDS[3], Duration.ofMillis(2));
-
-    ConcurrentHashMap<String, IncomingMessageEnvelope> recvEventsInputStartpointSpecific = new ConcurrentHashMap<>();
-    ConcurrentHashMap<String, IncomingMessageEnvelope> recvEventsInputStartpointTimestamp = new ConcurrentHashMap<>();
-    ConcurrentHashMap<String, IncomingMessageEnvelope> recvEventsInputStartpointOldest = new ConcurrentHashMap<>();
-    ConcurrentHashMap<String, IncomingMessageEnvelope> recvEventsInputStartpointUpcoming = new ConcurrentHashMap<>();
-
-    CoordinatorStreamStore coordinatorStreamStore = createCoordinatorStreamStore(applicationConfig1);
-    coordinatorStreamStore.init();
-    StartpointManager startpointManager = new StartpointManager(coordinatorStreamStore);
-    startpointManager.start();
-
-    StartpointSpecific startpointSpecific = new StartpointSpecific(String.valueOf(sentEvents1.get(100).offset()));
-    writeStartpoints(startpointManager, inputKafkaTopic1, ZK_TEST_PARTITION_COUNT, startpointSpecific);
-
-    StartpointTimestamp startpointTimestamp = new StartpointTimestamp(sentEvents2.get(150).timestamp());
-    writeStartpoints(startpointManager, inputKafkaTopic2, ZK_TEST_PARTITION_COUNT, startpointTimestamp);
-
-    StartpointOldest startpointOldest = new StartpointOldest();
-    writeStartpoints(startpointManager, inputKafkaTopic3, ZK_TEST_PARTITION_COUNT, startpointOldest);
-
-    StartpointUpcoming startpointUpcoming = new StartpointUpcoming();
-    writeStartpoints(startpointManager, inputKafkaTopic4, ZK_TEST_PARTITION_COUNT, startpointUpcoming);
-
-    startpointManager.stop();
-    coordinatorStreamStore.close();
-
-    TestTaskApplication.TaskApplicationCallback processedCallback = (IncomingMessageEnvelope ime, TaskCallback callback) -> {
-      try {
-        String streamName = ime.getSystemStreamPartition().getStream();
-        TestStreamApplication.TestKafkaEvent testKafkaEvent =
-            TestStreamApplication.TestKafkaEvent.fromString((String) ime.getMessage());
-        String eventIndex = testKafkaEvent.getEventData();
-        if (inputKafkaTopic1.equals(streamName)) {
-          recvEventsInputStartpointSpecific.put(eventIndex, ime);
-        } else if (inputKafkaTopic2.equals(streamName)) {
-          recvEventsInputStartpointTimestamp.put(eventIndex, ime);
-        } else if (inputKafkaTopic3.equals(streamName)) {
-          recvEventsInputStartpointOldest.put(eventIndex, ime);
-        } else if (inputKafkaTopic4.equals(streamName)) {
-          recvEventsInputStartpointUpcoming.put(eventIndex, ime);
-        } else {
-          throw new RuntimeException("Unexpected input stream: " + streamName);
-        }
-        callback.complete();
-      } catch (Exception ex) {
-        callback.failure(ex);
-      }
-    };
-
-    // Create StreamApplication from configuration.
-    CountDownLatch processedMessagesLatchStartpointSpecific = new CountDownLatch(100); // Just fetch a few messages
-    CountDownLatch processedMessagesLatchStartpointTimestamp = new CountDownLatch(100); // Just fetch a few messages
-    CountDownLatch processedMessagesLatchStartpointOldest = new CountDownLatch(NUM_KAFKA_EVENTS); // Fetch all since consuming from oldest
-    CountDownLatch processedMessagesLatchStartpointUpcoming = new CountDownLatch(5); // Expecting none, so just attempt a small number of fetches.
-
-    CountDownLatch shutdownLatchStartpointSpecific = new CountDownLatch(1);
-    CountDownLatch shutdownLatchStartpointTimestamp = new CountDownLatch(1);
-    CountDownLatch shutdownLatchStartpointOldest = new CountDownLatch(1);
-    CountDownLatch shutdownLatchStartpointUpcoming = new CountDownLatch(1);
-
-    TestTaskApplication testTaskApplicationStartpointSpecific =
-        new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic1, outputKafkaTopic, processedMessagesLatchStartpointSpecific,
-            shutdownLatchStartpointSpecific, processedCallback);
-    TestTaskApplication testTaskApplicationStartpointTimestamp =
-        new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic2, outputKafkaTopic, processedMessagesLatchStartpointTimestamp,
-            shutdownLatchStartpointTimestamp, processedCallback);
-    TestTaskApplication testTaskApplicationStartpointOldest =
-        new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic3, outputKafkaTopic, processedMessagesLatchStartpointOldest,
-            shutdownLatchStartpointOldest, processedCallback);
-    TestTaskApplication testTaskApplicationStartpointUpcoming =
-        new TestTaskApplication(TEST_SYSTEM, inputKafkaTopic4, outputKafkaTopic, processedMessagesLatchStartpointUpcoming,
-            shutdownLatchStartpointUpcoming, processedCallback);
-
-    // Startpoint specific
-    ApplicationRunner appRunner = ApplicationRunners.getApplicationRunner(testTaskApplicationStartpointSpecific, applicationConfig1);
-    executeRun(appRunner, applicationConfig1);
-
-    assertTrue(processedMessagesLatchStartpointSpecific.await(1, TimeUnit.MINUTES));
-
-    appRunner.kill();
-    appRunner.waitForFinish();
-
-    assertTrue(shutdownLatchStartpointSpecific.await(1, TimeUnit.MINUTES));
-
-    Integer startpointSpecificOffset = Integer.valueOf(startpointSpecific.getSpecificOffset());
-    for (IncomingMessageEnvelope ime : recvEventsInputStartpointSpecific.values()) {
-      Integer eventOffset = Integer.valueOf(ime.getOffset());
-      String assertMsg = String.format("Expecting message offset: %d >= Startpoint specific offset: %d",
-          eventOffset, startpointSpecificOffset);
-      assertTrue(assertMsg, eventOffset >= startpointSpecificOffset);
-    }
-
-    // Startpoint timestamp
-    appRunner = ApplicationRunners.getApplicationRunner(testTaskApplicationStartpointTimestamp, applicationConfig2);
-
-    executeRun(appRunner, applicationConfig2);
-    assertTrue(processedMessagesLatchStartpointTimestamp.await(1, TimeUnit.MINUTES));
-
-    appRunner.kill();
-    appRunner.waitForFinish();
-
-    assertTrue(shutdownLatchStartpointTimestamp.await(1, TimeUnit.MINUTES));
-
-    for (IncomingMessageEnvelope ime : recvEventsInputStartpointTimestamp.values()) {
-      Integer eventOffset = Integer.valueOf(ime.getOffset());
-      assertNotEquals(0, ime.getEventTime()); // sanity check
-      String assertMsg = String.format("Expecting message timestamp: %d >= Startpoint timestamp: %d",
-          ime.getEventTime(), startpointTimestamp.getTimestampOffset());
-      assertTrue(assertMsg, ime.getEventTime() >= startpointTimestamp.getTimestampOffset());
-    }
-
-    // Startpoint oldest
-    appRunner = ApplicationRunners.getApplicationRunner(testTaskApplicationStartpointOldest, applicationConfig3);
-    executeRun(appRunner, applicationConfig3);
-
-    assertTrue(processedMessagesLatchStartpointOldest.await(1, TimeUnit.MINUTES));
-
-    appRunner.kill();
-    appRunner.waitForFinish();
-
-    assertTrue(shutdownLatchStartpointOldest.await(1, TimeUnit.MINUTES));
-
-    assertEquals("Expecting to have processed all the events", NUM_KAFKA_EVENTS, recvEventsInputStartpointOldest.size());
-
-    // Startpoint upcoming
-    appRunner = ApplicationRunners.getApplicationRunner(testTaskApplicationStartpointUpcoming, applicationConfig4);
-    executeRun(appRunner, applicationConfig4);
-
-    assertFalse("Expecting to timeout and not process any old messages.", processedMessagesLatchStartpointUpcoming.await(15, TimeUnit.SECONDS));
-    assertEquals("Expecting not to process any old messages.", 0, recvEventsInputStartpointUpcoming.size());
-
-    appRunner.kill();
-    appRunner.waitForFinish();
-
-    assertTrue(shutdownLatchStartpointUpcoming.await(1, TimeUnit.MINUTES));
   }
 
   /**
@@ -1474,15 +1261,5 @@ public class TestZkLocalApplicationRunner extends IntegrationTestHarness {
         containerModel.getTasks().forEach((taskName, taskModel) -> ssps.addAll(taskModel.getSystemStreamPartitions()));
       });
     return ssps;
-  }
-
-  private static CoordinatorStreamStore createCoordinatorStreamStore(Config applicationConfig) {
-    SystemStream coordinatorSystemStream = CoordinatorStreamUtil.getCoordinatorSystemStream(applicationConfig);
-    SystemAdmins systemAdmins = new SystemAdmins(applicationConfig);
-    SystemAdmin coordinatorSystemAdmin = systemAdmins.getSystemAdmin(coordinatorSystemStream.getSystem());
-    coordinatorSystemAdmin.start();
-    CoordinatorStreamUtil.createCoordinatorStream(coordinatorSystemStream, coordinatorSystemAdmin);
-    coordinatorSystemAdmin.stop();
-    return new CoordinatorStreamStore(applicationConfig, new NoOpMetricsRegistry());
   }
 }
